requirements.txt
google_api_python_client==1.7.2
google_auth_oauthlib==0.4.1
instructor==1.7.0
openai==1.57.0
protobuf==5.29.1
pydantic==2.10.3
python-dotenv==1.0.1
